+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/acceleration_creation/create/create_acceleration_header.tsx
@@ -12,7 +12,7 @@ import {
   EuiTitle,
 } from '@elastic/eui';
 import React from 'react';
-import { OPENSEARCH_ACC_DOCUMENTATION_URL } from '../../../../components/constants';
+import { OPENSEARCH_ACC_DOCUMENTATION_URL } from '../../../constants';
 
 export const CreateAccelerationHeader = () => {
   return (
