+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/acceleration_creation/selectors/define_index_options.test.tsx
