+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/acceleration_creation/selectors/index_advanced_settings.test.tsx
