+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/acceleration_creation/selectors/index_setting_options.tsx
