+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/acceleration_creation/selectors/selector_helpers/load_objects.tsx
